feat: add tool to resize terminal size
Inspired by the Archlinux Wiki. The Arch Wiki does not provide a
copyrights page, it is therefore hard to credit the contributors

Recent commit c6de5e8046d4965fde910ee3b0c9c0709899609c is 7 years old,
with reference to $wgRightsPage. Individual contributors can't be
referenced but there is credit to the specific page in the Arch wiki.

https://gitlab.archlinux.org/archlinux/infrastructure/-/blob/master/roles/archwiki/templates/LocalSettings.php.j2