Fix: Prevent error return code on void includes
The include statement will return an error code 20 if no data is included.

Their is three ways to fix this:
   - boilerplate jinja2 around each includes
   - include a dummy file with at least a function that do nothing
   - a function that does nothing in the included files

I chose the third as I found that signaling to the user the effects of its
pillar configuration is a good thing.

I also refined the jinja2 around the copy-all include to take care of more
specific setups where an user could have disabled all dotfiles but is enabling
some specifically.